Indirect generation of unidict.utf8 dictionary
Now not use unidic directory for dictionary
